refactor(admin): move src/lib/utilities files into lib/utils + lib/cms (audit C7)
git mv of the 16 leftover Payload-CMS-template utils: 9 generic ->
src/lib/utils/, 7 CMS-domain -> new src/lib/cms/ (+ their 2 tests to the
matching home). Fixes the moved files' cross-home imports:
- cms/{getDocument,getGlobals,getRedirects}: ./revealui-singleton ->
  @/lib/utils/revealui-singleton
- utils/__tests__/brandEnvFallback.test.ts: ../mergeOpenGraph ->
  @/lib/cms/mergeOpenGraph (test exercises both homes — mergeOpenGraph in
  cms, siteBranding in utils)
- utils/revealui-singleton.ts: JSDoc example path -> @/lib/utils
External importers rewritten in the next commit.
