+++ apps/admin/src/lib/cms/getRedirects.ts
@@ -1,7 +1,7 @@
 import type { Redirect } from '@revealui/core/types/admin';
 import { unstable_cache } from 'next/cache';
 import { asDocuments } from '@/lib/utils/type-guards';
-import { getRevealUIInstance } from './revealui-singleton';
+import { getRevealUIInstance } from '@/lib/utils/revealui-singleton';
 
 export async function getRedirects(depth = 1): Promise<Redirect[]> {
   const revealui = await getRevealUIInstance();
